+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 1.0.1 - 01-03-2023
+
+### Fixed
+
+- Fixed issue that, when using a regex, the prompt send to GPT was not vissible in the webview panel.
+- Fixed bug that the `selectedText` was given to the `getCompletionResult` function, but never used.
+- Removed timeout that extension did not load until VSCode was fully loaded.
+
+### Changes
+
+- Added a `Copy` button to the webview panel.
+- Render the original prompt as a `blockquote`.
+- Added some styling to the HTML webview panel.
+- Added the option to cancel the API request to OpenAI with a `cancel` button on the progress bar.
+
 ## 1.0.0 - 26-02-2023
 
 1.0.0 release of: `The Assistent`.